+++ tests/urls.py
@@ -6,7 +6,7 @@ from django.conf.urls import url
 from django_sysinfo.views import echo, sysinfo, version
 
 urlpatterns = (
-    url('sys/info/$', sysinfo, name='sys-info'),
-    url('sys/version/(?P<name>.*)/$', version, name='sys-version'),
-    url('sys/echo/(?P<value>.*)/$', echo, name='sys-echo'),
+    url("sys/info/$", sysinfo, name="sys-info"),
+    url("sys/version/(?P<name>.*)/$", version, name="sys-version"),
+    url("sys/echo/(?P<value>.*)/$", echo, name="sys-echo"),
 )